feat(ui): added register page. Closes #416
