--- apps/knowii/src/Components/Checkbox.tsx
@@ -1,12 +1,0 @@
-import classNames from 'classnames';
-import React from 'react';
-
-export default function Checkbox(props: React.DetailedHTMLProps<React.InputHTMLAttributes<HTMLInputElement>, HTMLInputElement>) {
-  return (
-    <input
-      type="checkbox"
-      {...props}
-      className={classNames('rounded border-gray-300 text-indigo-600 shadow-sm focus:ring-indigo-500', props.className)}
-    />
-  );
-}
