+++ apps/knowii/src/Pages/PrivacyPolicy.tsx
@@ -1,0 +1,12 @@
+import FooterGuest from '@/Components/FooterGuest';
+import { Head } from '@inertiajs/react';
+
+export default function TermsOfService() {
+  return (
+    <>
+      <Head title="Privacy policy" />
+      <h1 className="mt-16 text-center w-full">Coming soon...</h1>
+      <FooterGuest />
+    </>
+  );
+}
